requirements.txt
mtcnn
gunicorn
mediapipe
flask==2.2.2
Pillow==9.4.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.2.2
google-cloud-vision
opencv-python-headless
